requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
arxiv==2.1.3
asgiref==3.8.1
attrs==24.2.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.31.0
azure-identity==1.17.1
azure-search-documents==11.5.1
azure-storage-blob==12.23.0
backoff==2.2.1
beautifulsoup4==4.12.3
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.0
cryptography==43.0.1
cssselect2==0.7.0
cycler==0.12.1
dataclasses-json==0.6.7
deepdiff==8.0.1
Deprecated==1.2.14
distro==1.9.0
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
effdet==0.4.1
emoji==2.12.1
et-xmlfile==1.1.0
feedparser==6.0.11
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.65.0
gpt-researcher==0.9.6
greenlet==3.1.1
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
html5lib==1.1
htmldocx==0.0.6
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.25.0
humanfriendly==10.0
idna==3.10
iopath==0.1.10
isodate==0.6.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
json5==0.9.25
json_repair==0.30.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain==0.2.16
langchain-community==0.2.17
langchain-core==0.2.41
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langdetect==1.0.9
langgraph==0.2.32
langgraph-checkpoint==2.0.0
langsmith==0.1.121
layoutparser==0.3.4
loguru==0.7.2
lxml==5.3.0
lxml_html_clean==0.3.1
Markdown==3.7
markdown2==2.5.1
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
md2pdf==1.0.1
mistune==3.0.2
mpmath==1.3.0
msal==1.31.0
msal-extensions==1.2.0
msgpack==1.1.0
msrest==0.7.1
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx==1.16.2
onnxruntime==1.19.2
openai==1.46.0
opencv-python==4.10.0.84
openpyxl==3.1.5
orderly-set==5.2.2
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pgvector==0.3.4
pi_heif==0.18.0
pikepdf==9.2.1
pillow==10.4.0
playwright==1.48.0
portalocker==2.10.1
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.0.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydyf==0.11.0
pyee==12.0.0
PyJWT==2.9.0
PyMuPDF==1.24.11
pypandoc==1.13
pyparsing==3.1.4
pypdf==5.0.0
pypdfium2==4.30.0
pyphen==0.16.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
python-oxmsg==0.0.1
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
rapidfuzz==3.9.7
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
safetensors==0.4.5
scipy==1.14.1
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
sqlparse==0.5.1
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.7.0
timm==1.0.9
tinycss2==1.3.0
tokenizers==0.19.1
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.44.2
triton==3.0.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
unstructured==0.16.0
unstructured-client==0.25.9
unstructured-inference==0.7.36
unstructured.pytesseract==0.3.13
urllib3==2.2.3
uvicorn==0.31.0
uvloop==0.20.0
watchfiles==0.24.0
weasyprint==62.3
webencodings==0.5.1
websockets==13.1
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.2.0
yarl==1.11.1
zopfli==0.2.3
